// src/app/zdslogic-ui-base/core/models/profile-entity.model.ts
export class ProfileEntity {

	id: string;
	userId: string;
	userName: string;
	contactId: string;
	ownerId: string;
	createdAt: string;
	updatedAt: string;
	fullName: string;
	firstName: string;
	lastName: string;
	gender: string;
	birthdate: string;
	title: string;
	headline: string;
	addressCity: string;
	addressState: string;
	addressCountry: string;
	addressZip: string;
	company: string;
	industry: string;
	imageURL: string;
	linkedin: string;
	facebook: string;
	skype: string;
	twitter: string;
	notes: string;
	presenceImageUrl: string;
	email: string;
	bio: string;
	enabled: string;
}