// src/app/zdslogic-ui-plugins/contacts/core/interfaces/contact.model.ts
export interface Contact {

    id: number;
    userId: number;
    userName: string;
	ownerId: number;
	createdAt: string;
	updatedAt: string;
	fullName: string;
    firstName: string;
    lastName: string;
	gender: string;
	birthdate: string;
    title: string;
    company: string;
    imageURL: string;
	linkedin: string;
	facebook: string;
    skype: string;
    twitter: string;
    notes: string;
    presenceStatus: number;
    presenceImageUrl: string;
	email: string;
	bio: string;
    enabled: string;
}